requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
redis>=5.0.1
boto3>=1.34.0
openai>=1.40.0
deepgram-sdk>=3.2.7
elevenlabs>=1.1.0
telnyx>=2.0.0
python-socketio>=5.11.0
firebase-admin>=6.5.0
twilio>=9.0.0
sendgrid>=6.11.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
structlog>=24.1.0
python-multipart>=0.0.9
alembic>=1.13.1
httpx>=0.27.0
